Create script.py
Cron is a tool used for scheduling tasks in Unix/Linux operating systems. Here's how you can set up a Cron job to run your Python script daily:

Open your terminal and type the command crontab -e. This will open the Cron configuration file in your default editor.

Add a new line to the file with the following format:

This line specifies the schedule for the task to be run. The first five fields represent the minute, hour, day of the month, month, and day of the week, respectively. The * character means that the task should be run every minute, every hour, every day of the month, every month, and every day of the week. The sixth field is the command to be executed. In this case, it is the path to the Python executable followed by the path to your Python script.

Save and exit the file.
The above configuration will run your Python script every minute of every day. If you want to run your script at a specific time, you can modify the first five fields to specify the minute, hour, day of the month, month, and day of the week when you want the script to run.

For example, if you want to run the script every day at 6:00 AM, you can modify the configuration as follows:

This configuration will run your Python script at minute 0 of hour 6 every day. You can modify the configuration further to suit your specific needs.